+++ internal/orchestrator/engine.go
@@ -6,38 +6,12 @@ import (
 	"encoding/hex"
 	"fmt"
 	"log"
-	"sync"
 	"time"
 
 	"github.com/rocketship-ai/rocketship/internal/api/generated"
 	"github.com/rocketship-ai/rocketship/internal/dsl"
 	"go.temporal.io/sdk/client"
 )
-
-type Engine struct {
-	generated.UnimplementedEngineServer
-	temporal client.Client
-	runs     map[string]*RunInfo
-	mu       sync.RWMutex
-}
-type RunInfo struct {
-	ID        string
-	Name      string
-	Status    string
-	StartedAt time.Time
-	EndedAt   time.Time
-	Tests     map[string]*TestInfo // Test's WorkflowID : TestInfo
-	Logs      []string
-}
-
-type TestInfo struct {
-	WorkflowID string
-	Name       string
-	Status     string
-	StartedAt  time.Time
-	EndedAt    time.Time
-	RunID      string
-}
 
 func NewEngine(c client.Client) *Engine {
 	return &Engine{
